// src/app/types/heroes.ts
export const HEROES: Heroes[] = [
  "dwarf-f",
  "dwarf-m",
  "elf-f",
  "elf-m",
  "knight-f",
  "knight-m",
  "lizard-f",
  "lizard-m",
  "wizzard-f",
  "wizzard-m",
];

export type Heroes =
  | "dwarf-f"
  | "dwarf-m"
  | "elf-f"
  | "elf-m"
  | "knight-f"
  | "knight-m"
  | "lizard-f"
  | "lizard-m"
  | "wizzard-f"
  | "wizzard-m";